OP-931: adds -Wextra compiler option for the flight code, and makes the bazillion code changes required to make the flight code compile again. Needs careful review, particularly all the fixes for the signed vs unsigned comparisons.
+review OPReview-459

+++ flight/pios/inc/pios_rfm22b_priv.h
@@ -749,7 +749,7 @@ struct pios_rfm22b_dev {
 #ifdef PIOS_INCLUDE_RFM22B_RCVR
     // The PPM channel values
     uint16_t ppm_channel[PIOS_RFM22B_RCVR_MAX_CHANNELS];
-    uint8_t ppm_supv_timer;
+    uint32_t ppm_supv_timer;
     bool ppm_fresh;
 #endif
 };
